+++ Codive/Application/CodiveApp.swift
@@ -28,24 +28,3 @@ struct CodiveApp: App {
         }
     }
 }
-
-/// 앱의 최상위 RootView
-struct AppRootView: View {
-    
-    @StateObject var appRouter: AppRouter
-    private let authDIContainer: AuthDIContainer
-    
-    init(appDIContainer: AppDIContainer) {
-        self._appRouter = StateObject(wrappedValue: appDIContainer.appRouter)
-        self.authDIContainer = appDIContainer.makeAuthDIContainer()
-    }
-    
-    var body: some View {
-        switch appRouter.currentAppState {
-        case .auth:
-            authDIContainer.makeAuthView()
-        case .main:
-            MainTabView()
-        }
-    }
-}
