+++ Codive/Features/Closet/Presentation/View/ClosetView.swift
@@ -1,0 +1,18 @@
+//
+//  ClosetView.swift
+//  Codive
+//
+//  Created by 황상환 on 9/24/25.
+//
+
+import SwiftUI
+
+struct ClosetView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    ClosetView()
+}
