+++ Codive/Features/Profile/Presentation/View/ProfileView.swift
@@ -1,0 +1,18 @@
+//
+//  ProfileView.swift
+//  Codive
+//
+//  Created by 황상환 on 9/24/25.
+//
+
+import SwiftUI
+
+struct ProfileView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    ProfileView()
+}
