--- Codive/Presentation/Feature/Add/ViewModel/AddViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  AddViewModel.swift
-//  Codive
-//
-//  Created by 황상환 on 9/22/25.
-//
-
-import Foundation
