--- Codive/Presentation/Feature/Closet/View/ClosetView.swift
@@ -1,18 +1,0 @@
-//
-//  ClosetView.swift
-//  Codive
-//
-//  Created by 황상환 on 9/22/25.
-//
-
-import SwiftUI
-
-struct ClosetView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    ClosetView()
-}
