--- Codive/Presentation/Feature/Closet/ViewModel/ClosetViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  ClosetViewModel.swift
-//  Codive
-//
-//  Created by 황상환 on 9/22/25.
-//
-
-import Foundation
