--- Codive/Presentation/Feature/Feed/ViewModel/FeedViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  FeedViewModel.swift
-//  Codive
-//
-//  Created by 황상환 on 9/22/25.
-//
-
-import Foundation
