--- Codive/Presentation/Feature/Home/View/HomeView.swift
@@ -1,18 +1,0 @@
-//
-//  HomeView.swift
-//  Codive
-//
-//  Created by 황상환 on 9/22/25.
-//
-
-import SwiftUI
-
-struct HomeView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    HomeView()
-}
