--- Codive/Presentation/Feature/Home/ViewModel/HomeViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  HomeViewModel.swift
-//  Codive
-//
-//  Created by 황상환 on 9/22/25.
-//
-
-import Foundation
