--- Codive/Presentation/Feature/Profile/ViewModel/ProfileViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  ProfileViewModel.swift
-//  Codive
-//
-//  Created by 황상환 on 9/22/25.
-//
-
-import Foundation
